pyproject.toml
[tool.poetry]
name = "myplaner"
version = "0.1.0"
description = ""
authors = ["Dmitry Grishin <kenpxrkasf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.109.2"}
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.27"
pip = "^24.0"
install = "^1.3.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "==4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
